QC progress, second level analyses

+++ pipeline/workflow/fake.py
@@ -6,6 +6,10 @@ import fasteners
 import json
 
 from shutil import copy, copyfileobj
+from pathlib import Path
+
+import re
+from pkg_resources import resource_filename as pkgr
 
 from nipype.interfaces.base import (
     isdefined, 
@@ -46,11 +50,12 @@ def _find(target, d):
 class FakeDerivativesDataSink(DerivativesDataSink):
     def __init__(self, 
             images, 
-            fmriprep_reportlets_dir, fmriprep_output_dir, 
+            output_dir, fmriprep_reportlets_dir, fmriprep_output_dir, 
             node_id, depends, 
             **inputs):
         super(FakeDerivativesDataSink, self).__init__(**inputs)
         
+        self.output_dir = output_dir
         self.fmriprep_reportlets_dir = fmriprep_reportlets_dir
         self.fmriprep_output_dir = fmriprep_output_dir
         self.node_id = node_id
@@ -67,6 +72,9 @@ class FakeDerivativesDataSink(DerivativesDataSink):
             if "anat" in self.inputs.in_file[0]:
                 out_path = out_path + "/" + "T1w"
         
+        if out_path is None:
+            out_path = ""
+        
         _, ext = _splitext(self.inputs.in_file[0])
         compress = ext == '.nii'
         if compress:
@@ -81,34 +89,42 @@ class FakeDerivativesDataSink(DerivativesDataSink):
             return runtime
         elif base_directory == self.fmriprep_reportlets_dir:
             # write to json
-            
             work_dir = op.dirname(base_directory)
-            json_data = {"id": "%s.%s" % (self.node_id, self.inputs.suffix), 
-                "depends": self.depends, "html": ""}
-            for i, fname in enumerate(self.inputs.in_file):
-                with open(fname, "r") as f:
-                    json_data["html"] += f.read()
+            json_id = "%s.%s" % (self.node_id, self.inputs.suffix)
+            json_id = re.sub(r'func_preproc_[^.]*', "func_preproc_wf", json_id)
+            json_data = {"id": json_id}
             
-            with fasteners.InterProcessLock(op.join(work_dir, "qc.lock")):
-                json_file = op.join(work_dir, "qc.json")
-                    
-                with open(json_file, "ab+") as f:
-                    f.seek(0, 2)       
-                    
-                    if f.tell() == 0:
-                        f.write(json.dumps([json_data]).encode())
-                    else:
-                        f.seek(-1, 2)
-                        f.truncate()
-                        f.write(','.encode())
-                        f.write(json.dumps(json_data).encode())
-                        f.write(']'.encode())  
+            out_path = op.join(out_path, "qualitycheck")
+            os.makedirs(op.join(self.output_dir, out_path), exist_ok = True)
+            
+            touch_fname = op.join(out_path, json_data["id"] + ext)
+            touch_path = op.join(self.output_dir, touch_fname)
+            
+            if not op.isfile(touch_path):
+                for i, fname in enumerate(self.inputs.in_file):
+                    copy(fname, touch_path)
+                    # with open(fname, "r") as f:
+                    #     json_data["html"] += f.read()
+                json_data["fname"] = op.join(op.basename(self.output_dir), touch_fname)
+                with fasteners.InterProcessLock(op.join(work_dir, "qc.lock")):
+                    json_file = op.join(work_dir, "qc.json")
+                    with open(json_file, "ab+") as f:
+                        f.seek(0, 2)       
+                        
+                        if f.tell() == 0:
+                            f.write(json.dumps([json_data]).encode())
+                        else:
+                            f.seek(-1, 2)
+                            f.truncate()
+                            f.write(','.encode())
+                            f.write(json.dumps(json_data).encode())
+                            f.write(']'.encode())  
+                Path(touch_path).touch()
+            html_path = op.join(op.dirname(self.output_dir), "index.html")
+            if not op.isfile(html_path):
+                copy(pkgr('pipeline', 'index.html'), html_path)
         else:
             # copy file to out_path
-            
-            if out_path is None:
-                out_path = ""
-                
             out_path = op.join(base_directory, out_path)
 
             os.makedirs(out_path, exist_ok = True)
